requirements.txt
dj-database-url==0.5.0
Django==2.2
django-crispy-forms==1.7.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
